pyproject.toml

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "devjoni-arenaprog"
version = "0.0.5"
authors = [
  { name="DEV Joni", email="solutions@devjoni.com" },
]
description = "Experiment control program for the arena"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
	"pyserial",
	"pillow",
	"devjoni-hosguibase",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
]

keywords = ["toolkit", "GUI", "tk", "gtk", "p3d"]

[project.urls]
"Homepage" = "https://github.com/devjonix/arenaprog"
"Bug Tracker" = "https://github.com/devjonix/arenaprog/issues"
